Move dbkeys to managers
