+++ lib/tool_shed/webapp/app.py
@@ -18,6 +18,7 @@ from galaxy.app import (
 from galaxy.config import configure_logging
 from galaxy.managers.api_keys import ApiKeyManager
 from galaxy.managers.citations import CitationsManager
+from galaxy.managers.dbkeys import GenomeBuilds
 from galaxy.managers.users import UserManager
 from galaxy.model.base import SharedModelMapping
 from galaxy.model.tags import CommunityTagHandler
@@ -27,7 +28,6 @@ from galaxy.quota import (
 )
 from galaxy.security import idencoding
 from galaxy.structured_app import BasicSharedApp
-from galaxy.util.dbkeys import GenomeBuilds
 from galaxy.web_stack import application_stack_instance
 from tool_shed.grids.repository_grid_filter_manager import RepositoryGridFilterManager
 from tool_shed.structured_app import ToolShedApp
